Basic script viewer implemented.

+++ tools/binary_diff.js
@@ -13,7 +13,7 @@ function readROM(which) {
       rom[which] = {};
       rom[which].u8 = new Uint8Array(e.target.result);
       rom[which].u16 = new Uint16Array(e.target.result);
-      rom[which].u32 = new Uint16Array(e.target.result);
+      rom[which].u32 = new Uint32Array(e.target.result);
     };
     fileReader.readAsArrayBuffer(file);
   };
